requirements.txt
weaviate-client
mistralai
pandas
llama-index